% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate_stop_times}
\alias{interpolate_stop_times}
\title{Interpolate the times vehicles actually passed each stop,
based on vehicle positions.}
\usage{
interpolate_stop_times(
  positions,
  static,
  extrapolate_speed_kmh = 25,
  dist_thresh_meters = 2000,
  off_route_max_proportion = 0.7,
  on_route_min_stops = 4
)
}
\description{
This works by modifying current_stop_sequence s.t. current_stop_sequence = x means the exact moment
we arrived at the stop
IN_TRANSIT_TO means have not arrived at stop yet, so we make the stop sequence smaller -
we want to find the last IN_TRANSIT_TO this stop (or, if there isn't one, the last STOPPED_AT
from an earlier stop). The spec is unclear on whether this means in transit to the stop or to the next,
but the GoTriangle GTFS anyways seems to use it to mean going to the stop.
INCOMING_AT is similar but closer to arrival
STOPPED_AT means we have arrived at the stop in the past, so add a little
We then interpolate an arrival time based on distance from the previous and to the next report,
so exact stop sequence spacing is not important, just ordering
}
