% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petrify.R
\name{petrify}
\alias{petrify}
\title{"Petrify" a day of GTFS-realtime data, i.e. turn it into a GTFS static feed.}
\usage{
petrify(rt, static, override_calendar = TRUE, ...)
}
\description{
This takes a data.table with vehicle positions, and a static {gtfstools} GTFS feed object
that it matches with, and returns a new static {gtfstools} GTFS feed with the trips that were
operated based on the realtime data.
}
\details{
The realtime data should be for a single service day.
}
