% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GtfsRtAccuracy-package.R,
%   R/petrifygtfsrt-package.R
\docType{package}
\name{petrifygtfsrt-package}
\alias{petrifygtfsrt-package}
\title{petrifygtfsrt: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).

What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

Authors:
\itemize{
  \item First Last \email{first.last@example.com}
}

}
\keyword{internal}
