% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_chunk}
\alias{read_chunk}
\title{Read a chunk of updates and positions}
\usage{
read_chunk(base_path, start_date, end_date, timezone)
}
\description{
The unit of analysis is the update. So we read n days of updates.
Particularly because the data are stored in UTC days (but also because
some agencies run service over midnight even if we converted to local
time), we may have updates for trips that run on the next day. So we read
n days of updates, but n+1 days of positions. This way, every update that
can match to a position will, and there is a many to one relationship between
updates and positions anyhow, so we're not duplicating any information.
We may read the same position in multiple chunks, but each update will get
read exactly once (i.e. once for each time it appears, so an update that
appears for 20 minutes will appear in 20 files).
}
\details{
Expects that the data are arranged in the format used by the UNC GTFS-realtime
archive, i.e. base_path/year/month/day/{vehicle_positions|trip_updates}/*,
with static gtfs in base_path/static/year-month-day.zip.
}
