% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_static_stop_times}
\alias{read_static_stop_times}
\title{Read all of the static GTFS into a big table of stop times,
with service_day_start indicating the date the data were retrieved.}
\usage{
read_static_stop_times(base_path)
}
\description{
Read all of the static GTFS into a big table of stop times,
with service_day_start indicating the date the data were retrieved.
}
